partially fill analytics profile page with data gotten from API

+++ src/application/MindGameApplication/app/src/main/java/anekdotas/mindgameapplication/network/AnalyticModel.kt
@@ -3,7 +3,7 @@ package anekdotas.mindgameapplication.network
 data class AnalyticModel (
     val totalTimeSpent: Int = 0,
     val correctAnswers : Int = 0,
-    val correctAnswersPercentage :Int = 0,
+    val correctAnswersPercentage : Double = 0.0,
     val longestStreak: Int =0,
     val longestStreakTopicId: Int = 0,
     val averageGameTime: Int = 0,
